requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.2.2
Django==2.2.2
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
psycopg2==2.8.2
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
typing-extensions==3.7.2
websockets==6.0
whitenoise==4.1.2
yarl==1.3.0
